+++ src/command_modules/azure-cli-cosmosdb/azure/cli/command_modules/cosmosdb/custom.py
@@ -3,6 +3,8 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
+from __future__ import print_function
+
 from knack.log import get_logger
 from knack.util import CLIError
 
@@ -186,6 +188,11 @@ def cli_cosmosdb_update(client,
 
     if enable_multiple_write_locations is None:
         enable_multiple_write_locations = existing.enable_multiple_write_locations
+    elif enable_multiple_write_locations != existing.enable_multiple_write_locations and \
+            enable_multiple_write_locations:
+        raise CLIError("Cannot convert account from single master to multi master")
+    elif enable_multiple_write_locations != existing.enable_multiple_write_locations:
+        print("This operation will take 24 hours to complete")
 
     params = DatabaseAccountCreateUpdateParameters(
         location=existing.location,