+++ src/command_modules/azure-cli-cosmosdb/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 
 0.2.6 
 +++++ 
-* Added support for updating account to multi-master
+* Added support for updating account from multi-master to single-master
 
 0.2.5
 +++++